Add safety checks to publish script

+++ bin/publish.sh
@@ -1,12 +1,24 @@
 #! /bin/bash
 
 VERSION=`node bin/getVersion.js  2>/dev/null`
+
+if [ -z "$1" ]; then
+    echo "Build type not specified"
+    echo
+    exit 1
+fi
+
+if [[ $VERSION = *"-"* ]] && [ $1 = "prod" ]; then
+    echo "Cannot publish a beta version as prod"
+    echo
+    exit 1
+fi
+
 echo
 echo "Publishing version $VERSION"
 echo
 
-
-npm run build:prod
+npm run build:$1
 
 cd dist/parent
 npm publish
@@ -20,6 +32,13 @@ cd ../react
 npm publish
 cd ../vue
 npm publish
+
+if [ $1 = "beta" ] 
+then
+  exit 0
+fi
+
+echo "Updating GitHub build"
 
 cd ../..
 rm -v iframe-resizer.zip
@@ -35,4 +54,5 @@ git pull
 git push
 git push --tags
 
+echo "Updating iframe-resizer.com"
 cp -v js/* ../docs/public/js
